# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "golang作品",
        "title": "Gin-Backend",
        "termsOfService": "github",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/category/add": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Category AddCategory",
                "parameters": [
                    {
                        "description": "Add",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CategoryResult"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/category/delete": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Category DeleteCategory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Delete Category",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/category/edit": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Category EditCategory",
                "parameters": [
                    {
                        "description": "Edit Category",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Category"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/category/info": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Category CategoryInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Info Query id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/category/list4backend": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Category CategoryListForBackend",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Category": {
            "type": "object",
            "properties": {
                "categoryId": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "isDeleted": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "parentId": {
                    "type": "string"
                }
            }
        },
        "model.CategoryResult": {
            "type": "object",
            "properties": {
                "c1CategoryId": {
                    "type": "string"
                },
                "c1Desc": {
                    "type": "string"
                },
                "c1Name": {
                    "type": "string"
                },
                "c1Order": {
                    "type": "integer"
                },
                "c1ParentId": {
                    "type": "string"
                },
                "c2CategoryId": {
                    "type": "string"
                },
                "c2Name": {
                    "type": "string"
                },
                "c2Order": {
                    "type": "integer"
                },
                "c2ParentId": {
                    "type": "string"
                },
                "c3CategoryId": {
                    "type": "string"
                },
                "c3Name": {
                    "type": "string"
                },
                "c3Order": {
                    "type": "integer"
                },
                "c3ParentId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isDeleted": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                }
            }
        },
...
    }
}